+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     * Fix free when using a reboot action
     * Added `datetime_format` as a configuration in srd.conf
     * Action `log` now supports `user` which sets the owner of the logfile (only on creation)
+    * Added `%downtime` placeholder for the downtime in days, hours, minutes and seconds
 
 * 0.0.4
     * Fix some memory not getting deallocated